% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hint_stud_table.R
\name{err.sol.table.for.ps}
\alias{err.sol.table.for.ps}
\title{Create an err.sol.table from a single problem set}
\usage{
err.sol.table.for.ps(err.df, ps.name, rps.dir = getwd(),
  rps = load.rps(file.path(rps.dir, ps.name)))
}
\description{
see err.sol.table for a basic description
}
