% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hint_stud_table.R
\name{err.sol.table}
\alias{err.sol.table}
\title{Create an err.sol.table as input for hint.stud.table}
\usage{
err.sol.table(err.df, rps.dir = getwd())
}
\arguments{
\item{err.df}{If res is the the returned list from a call to analyse.subs() it is the elemet res$err.df}

\item{rps.dir}{The directory in which all problem set rps files can be found of the problem sets that have been analysed by analyse.subs.}
}
\description{
Create an err.sol.table as input for hint.stud.table
}
