% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_sub.R
\name{grade.subs}
\alias{grade.subs}
\title{Combine for each student the points from all problem sets and create
csv files with the total points.}
\usage{
grade.subs(sub.li, grade.dir = paste0(base.dir, "/grades"),
  base.dir = getwd())
}
\description{
See README.md for usage
}
