% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hint_stud_table.R
\name{hint.stud.table}
\alias{hint.stud.table}
\title{Creates a table with hint.stud templates from an err.sol.table}
\usage{
hint.stud.table(es, min.users = 1, default.filters = TRUE)
}
\arguments{
\item{es}{a data frame returned by err.sol.table, err.sol.table.for.ps or err.sol.table.for.chunk}

\item{min.users}{only keep wrong calls that at least min.users made}

\item{default.filters}{if TRUE remove some rows that are unlikely to be used as custom hints (e.g. if in assignment both the assigned variable and the rhs differ).}
}
\description{
Creates a table with hint.stud templates from an err.sol.table
}
