% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sub.R
\name{load.subs}
\alias{load.subs}
\title{Load all submissions from a directory}
\usage{
load.subs(base.dir = NULL, files = NULL, stud.name.fun = NULL,
  sub.dir = file.path(base.dir, "sub"), warn = TRUE)
}
\description{
See README.md for usage
}
