% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_error_hint.R, R/analyse_sub_2019.R
\name{unpack.big.zips}
\alias{unpack.big.zips}
\alias{unpack.big.zips}
\title{Takes assignment ZIPs with all students' solutions
and unpacks them into separate folders for each assignment}
\usage{
unpack.big.zips(base.dir, prefix = "", postfix = ".zip")

unpack.big.zips(base.dir, prefix = "", postfix = ".zip")
}
\description{
Takes assignment ZIPs with all students' solutions
and unpacks them into separate folders for each assignment

Takes assignment ZIPs with all students' solutions
and unpacks them into separate folders for each assignment
}
