% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle.R
\name{unpack.moodle.sub.zips}
\alias{unpack.moodle.sub.zips}
\title{Takes assignment ZIPs with all students' solutions
and unpacks them into separate folders for each assignment}
\usage{
unpack.moodle.sub.zips(zip.dir = "moodle_zip", sub.dir = "sub",
  prefix = "", postfix = ".zip")
}
\arguments{
\item{zip.dir}{directory with big ZIP files from Moodle. Each ZIP file contains all submissions of one problem set}

\item{sub.dir}{directory into which sub files shall be extracted}
}
\description{
Takes assignment ZIPs with all students' solutions
and unpacks them into separate folders for each assignment
}
