% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_sub.R
\name{analyse.subs}
\alias{analyse.subs}
\title{Summarize number of errors and hints by
looking at the submission logs}
\usage{
analyse.subs(sub.li, base.dir = getwd(), rps.dir = file.path(base.dir,
  "ps"), no.summary = FALSE)
}
\description{
Summarize number of errors and hints by
looking at the submission logs
}
