% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle.R
\name{load.moodle.subs}
\alias{load.moodle.subs}
\title{Load submission files that follow
Moodle's naming convention}
\usage{
load.moodle.subs(base.dir = getwd(), sub.dir = file.path(base.dir, "sub"),
  stud.name.fun = moodle.stud.name.fun, ...)
}
\description{
Load submission files that follow
Moodle's naming convention
}
