% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sub.R
\name{load.subs}
\alias{load.subs}
\title{Load all submissions from a directory}
\usage{
load.subs(sub.dir = "sub", files = NULL, stud.name.fun = NULL,
  warn = TRUE)
}
\arguments{
\item{sub.dir}{The directory in which all submission files can be found.}
}
\description{
See README.md for usage
}
