% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle.R
\name{unpack.moodle.sub.zips}
\alias{unpack.moodle.sub.zips}
\title{Takes assignment ZIPs with all students' solutions
and unpacks them into separate folders for each assignment}
\usage{
unpack.moodle.sub.zips(base.dir, zip.dir = file.path(base.dir, "moodle_zip"),
  sub.dir = file.path(base.dir, "sub"), prefix = "", postfix = ".zip")
}
\description{
Takes assignment ZIPs with all students' solutions
and unpacks them into separate folders for each assignment
}
